пр14 сделана авторизация, центральная обработка ошибок, валидация тела входящего запроса v1.0

+++ controllers/cards.js
@@ -1,10 +1,11 @@
 const cardModel = require("../models/card");
-const { serverError, typeError } = require("../utils/errors")
+const { serverError, typeError } = require("../middlewares/handleErrors")
 
 const {
   HTTP_STATUS_OK,                   // 200
-  HTTP_STATUS_CREATED               // 201
-} = require('../utils/constantsError')
+  HTTP_STATUS_CREATED,             // 201
+  HTTP_STATUS_BAD_REQUEST           // 400
+} = require('../utils/constantsStatusCode')
 
 function getAllCards(req, res) {
   return cardModel
@@ -12,9 +13,7 @@ function getAllCards(req, res) {
     .then((cards) => {
       return res.status(HTTP_STATUS_OK).send(cards);
     })
-    .catch((err) => {
-      serverError(err, res)
-    });
+    .catch(next);
 }
 
 function createCard(req, res) {
@@ -26,63 +25,58 @@ function createCard(req, res) {
     .then((card) => {
       return res.status(HTTP_STATUS_CREATED).send(card);
     })
-    .catch((err) => {
-      typeError(err, res)
-    });
+    .catch(next);
 }
 
 function deleteCard(req, res) {
   console.log(req.params);
 
   const { cardId } = req.params;
-
-  return cardModel
-    .findByIdAndDelete(cardId)
-    .orFail()
-    .then((card) => {
-      return res.status(HTTP_STATUS_OK).send(card);
-    })
-    .catch((err) => {
-      typeError(err, res)
-    });
+  try {
+    if(cardId === req.user._id) {
+      return cardModel
+        .findByIdAndDelete(cardId)
+        .orFail()
+        .then((card) => {
+          return res.status(HTTP_STATUS_OK).send(card);
+        })
+        .catch(next);
+    }
+    return Promise.reject(new Error('Чужая карточка!'))
+  } catch (err) {
+    next(err)
+  }
 }
 
-function likeCard(req, res) {
-
+function toggleLikeCard(req, res, methodObj) {
   const { cardId } = req.params;
   return cardModel
     .findByIdAndUpdate(
       cardId,
-      { $addToSet: { likes: req.user._id } }, // добавить _id в массив, если его там нет
+      methodObj, // добавить _id в массив, если его там нет
       { new: true },
     )
     .orFail()
     .then((card) => {
       return res.status(HTTP_STATUS_OK).send(card);
     })
-    .catch((err) => {
-      typeError(err, res)
-    });
+    .catch(next);
 }
 
-function deleteLikeCard(req, res) {
-
-  const { cardId } = req.params;
-
-  return cardModel
-    .findByIdAndUpdate(
-      cardId,
-      { $pull: { likes: req.user._id } }, // убрать _id из массива
-      { new: true },
-    )
-    .orFail()
-    .then((card) => {
-      return res.status(HTTP_STATUS_OK).send(card);
-    })
-    .catch((err) => {
-      typeError(err, res)
-    });
+function togglelikeCardDecorator(func) {
+  return function(req, res) {
+    if (req.method === 'put') {
+      return func(req, res, { $addToSet: { likes: req.user._id } }) // добавить _id в массив, если его там нет
+    }
+    if (req.method === 'delete') {
+      return func(req, res, { $pull: { likes: req.user._id } })   // убрать _id из массива
+    }
+  }
 }
+
+const likeCard = togglelikeCardDecorator(toggleLikeCard)
+const deleteLikeCard = togglelikeCardDecorator(toggleLikeCard)
+
 
 module.exports = {
   getAllCards,
